feat: Implement more Agent functionalities (#10)
* feat(fusion-ai): add tencent map mcp client example

Add example demonstrating how to use RMCP client to interact with Tencent Map MCP services
Enable additional rmcp features and dependencies required for the client
Include README with setup instructions and usage examples

* refactor(task): rename task request/response types and improve shutdown handling

- Rename TaskRequest to AcquireTaskRequest and TaskResponse to AcquireTaskResponse for clarity
- Replace broadcast channel with dedicated shutdown mechanism using mea::shutdown
- Simplify connection manager and task scheduler by extracting runner logic
- Add example shutdown implementation for reference

* refactor(task): simplify task execution and process management

- Replace UUID process_id with u32 pid for simpler process tracking
- Remove retry logic from TaskExecutor and rely on process manager
- Consolidate TaskExecutionError into enum and simplify error handling
- Update ResourceLimits structure with better defaults and documentation
- Rename JobConfig to TaskConfig for clarity
- Streamline process event handling and cleanup

* refactor(process_manager): improve process management and error handling

- Replace mpsc with broadcast for process events
- Use tokio::process::Child and Arc<Mutex> for process handles
- Change process_id from u32 to Uuid for better tracking
- Simplify cleanup and monitoring logic
- Improve error handling in task execution

* refactor(agent): restructure task execution and process management

- Replace multiple task executors with a single executor
- Simplify process event handling and status reporting
- Improve process killing with proper signal handling
- Add platform-specific process management
- Update task scheduler to use new executor model
- Clean up unused code and improve error handling

* refactor(agent): simplify async select statements by removing futures_util

docs: update todo.md to mark completed tasks

* refactor(task scheduling): improve task scheduling and process management

- Change max_concurrent_processes type from usize to u32 for consistency
- Add available_capacity method to ProcessManager
- Simplify task scheduling logic by removing unused state tracking
- Update task acquisition to use process manager's available capacity
- Add connection retry logic with timeout for websocket
- Mark concurrent control as completed in todo.md

* feat(agent): implement agent registration and websocket communication

refactor(core): update websocket protocol and message formats

docs: add requirements documentation and update architecture diagrams

* feat(task): implement task retry mechanism

- Add retry_count field to SchedTask model
- Implement task instance status handling with retry logic
- Add generate_retry_tasks service for failed tasks
- Update todo.md documentation for retry feature

* feat(server): implement heartbeat monitoring and logging improvements

- Add server and agent heartbeat timeout monitoring
- Implement structured logging configuration for both server and agent
- Update server model to track last heartbeat timestamp
- Add heartbeat response protocol structure
- Remove unused HeartbeatRequest import
- Update documentation and todo lists

+++ hetuflow/hetuflow-agent/src/setting.rs
@@ -117,7 +117,7 @@ pub struct ProcessConfig {
   pub process_timeout: Duration,
 
   /// 最大并发进程数，同时可以调度执行的任务数
-  pub max_concurrent_processes: usize,
+  pub max_concurrent_processes: u32,
 
   /// 启用资源监控
   pub enable_resource_monitoring: bool,
